Pet recyclerView일단 수정
그다음 할일은 null값 없애기

+++ app/src/main/java/com/example/project_petdoc/RegisterActivity.kt
@@ -21,7 +21,7 @@ class RegisterActivity : AppCompatActivity() {
                 putExtra("category", binding.editCategory.text.toString())
                 putExtra("name", binding.editName.text.toString())
                 putExtra("gender", binding.editS.text.toString())
-                putExtra("age", binding.editAge.text.toString().toInt()) // String을 Int로 변환
+                putExtra("age", binding.editAge.text.toString())
                 putExtra("hospital", binding.editHos.text.toString())
             }
             setResult(RESULT_OK, intent)
